Refactor components and get initial functionality
