+++ clients/admin-ui/src/features/privacy-requests/RequestFilters.tsx
@@ -10,16 +10,13 @@ import {
   useToast,
 } from "@fidesui/react";
 import MultiSelectDropdown from "common/dropdown/MultiSelectDropdown";
+import { CloseSolidIcon, DownloadSolidIcon, SearchLineIcon } from "common/Icon";
+import PIIToggle from "common/PIIToggle";
 import React, { useCallback, useMemo } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
 import { selectToken } from "../auth";
-import {
-  CloseSolidIcon,
-  DownloadSolidIcon,
-  SearchLineIcon,
-} from "../common/Icon";
-import PIIToggle from "../common/PIIToggle";
+import { SubjectRequestStatusMap } from "./constants";
 import {
   clearAllFilters,
   requestCSVDownload,
@@ -28,9 +25,8 @@ import {
   setRequestId,
   setRequestStatus,
   setRequestTo,
-} from "../privacy-requests/privacy-requests.slice";
-import { PrivacyRequestStatus } from "../privacy-requests/types";
-import { SubjectRequestStatusMap } from "./constants";
+} from "./privacy-requests.slice";
+import { PrivacyRequestStatus } from "./types";
 
 const useRequestFilters = () => {
   const filters = useSelector(selectPrivacyRequestFilters);
