+++ clients/admin-ui/src/features/subject-request/events-and-logs/EventError.tsx
