+++ clients/admin-ui/src/pages/index.tsx
@@ -5,8 +5,8 @@ import { LOGIN_ROUTE } from "../constants";
 import ProtectedRoute from "../features/auth/ProtectedRoute";
 import Head from "../features/common/Head";
 import NavBar from "../features/common/NavBar";
+import RequestFilters from "../features/privacy-requests/RequestFilters";
 import RequestTable from "../features/privacy-requests/RequestTable";
-import RequestFilters from "../features/subject-request/RequestFilters";
 
 const Home: NextPage = () => (
   <ProtectedRoute redirectUrl={LOGIN_ROUTE}>
